Fix compile warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@159357 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 9929d6f62065036751750e610a9cf00f83596338

+++ Online/Online/OnlineBase/src/MBM/Summary.cpp
@@ -76,12 +76,12 @@ int Summary::show(BufferMemory* dsc, bool show_states)   {
 
   ::printf("\n======================== MBM Bufer summary for buffer \"%s\" ========================\n\n",dsc->bm_name);
   ::sprintf(txt," Buffer \"%s\"",dsc->bm_name);
-  ::printf("%-26s  Events: Produced:%ld Seen:%ld Pending:%ld Max:%d\n",
-	   txt, long(ctr->tot_produced), long(ctr->tot_seen), ctr->i_events, ctr->p_emax);
+  ::printf("%-26s  Events: Produced:%ld Seen:%ld Pending:%ld Max:%ld\n",
+	   txt, long(ctr->tot_produced), long(ctr->tot_seen), long(ctr->i_events), long(ctr->p_emax));
   ::printf("%-26s  Space(kB):[Tot:%ld Free:%ld] Users:[Tot:%ld Max:%ld]\n\n",
 	   "",long((ctr->bm_size*ctr->bytes_p_Bit)/1024), 
 	   long((ctr->i_space*ctr->bytes_p_Bit)/1024), 
-	   ctr->i_users, ctr->p_umax);
+	   long(ctr->i_users), long(ctr->p_umax));
 
   bool first = true;
   Users users;
